Test: Introduce system-test for cobbler buildiso
This test is currently very basic and only checks for the existence
of the ISO. In the future we should also check for the validity of
the content of the ISO. This is not yet possible as the tool for this
(parti) is not yet available everywhere.

What is missing with this basic test is also the different
use-cases for the buildiso command: Airgapped, Offline and Online
media creation.